+++ examples/scatter/SimpsonChart.tsx
@@ -9,8 +9,8 @@ import {
     Typography,
     Legend,
     LegendItem,
-    Segment,
 } from '@chsk/core'
+import { Segment } from '@chsk/annotation'
 import { isScatterData, Scatter, ScatterPoints, Regression } from '@chsk/xy'
 import { generateRegressionData } from '../line/generators'
 import { MilestoneStory } from '../types'