+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 
 ### Changed
 
+-   [packages/core]: moved several symbol shapes to the annotation package
 -   [packages/core]: updated props on `View` and `BaseView` to better match
     conventions on other components
 -   [packages/core]: modified default styles for some components appearing